Refactor: extract computeCumulativeCaseCounts() to Utils module.
